remaining compilation fixes for random gcc versions

+++ libs9s/s9srpcreply.cpp
@@ -936,7 +936,7 @@ S9sRpcReply::printClusterListLong()
     }
    
     if (!options->isBatchRequested())
-        printf("Total: %u\n", theList.size());
+        printf("Total: %lu\n", (unsigned long int) theList.size());
 }
 
 /**
